+++ src/app/View/Components/C404.php
@@ -3,11 +3,14 @@
 namespace App\View\Components;
 
 use Roots\Acorn\View\Component;
+use function Roots\asset;
 
 class C404 extends Component
 {
+    public $bg;
     public function __construct()
     {
+        $this->bg = asset('images/not-found.png')->uri();
     }
 
     public function render()