+++ src/resources/assets/scripts/routes/history.js
@@ -1,11 +1,11 @@
 /* eslint-disable */
 const { render } = wp.element;
-import History from '../components/history.jsx';
+import History from '../components/History.jsx';
 /* eslint-enable */
 
 export default {
   init() {
-    render(<History />, document.getElementById('react-app'));
+    render(<History />, document.getElementById('react-history-app'));
   },
   finalize() {
     // JavaScript to be fired on all pages, after page specific JS is fired