+++ src/resources/views/components/lighthouse.blade.php
@@ -1,0 +1,1 @@
+<div id="react-lighthouse-app" data-lottie-light="{!! $light !!}" data-lottie-dark="{!! $dark !!}" style="translate3d(0,0,0)"></div>